[dyad] Updated Redis controller and fixed scrollbar - wrote 10 file(s), added StackExchange.Redis package(s)

+++ backend/WatchTower/WatchTower.Shared/Models/RedisLatencyData.cs
@@ -3,6 +3,6 @@ namespace WatchTower.Shared.Models
     public class RedisLatencyData
     {
         public string Time { get; set; }
-        public int LatencyMs { get; set; }
+        public double LatencyMs { get; set; }
     }
 }
